@@ -29,10 +29,11 @@
   "dependencies": {
     "babel-core": "^6.14.0",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-polyfill": "^6.13.0",
+    "babel-polyfill": "^6.23.0",
     "babel-preset-es2015": "^6.14.0",
     "babel-preset-react": "^6.11.1",
     "babel-preset-stage-0": "^6.5.0",
+    "babel-runtime": "^6.23.0",
     "bluebird": "^3.4.1",
     "express": "^4.15.2",
     "react": "^15.3.1",
@@ -42,6 +43,7 @@
     "babel-cli": "^6.24.0",
     "babel-eslint": "^6.1.2",
     "babel-loader": "^6.2.5",
+    "babel-plugin-transform-runtime": "^6.23.0",
     "babel-preset-es2017": "^6.22.0",
     "clean-webpack-plugin": "^0.1.10",
     "css-loader": "^0.24.0",
@@ -59,7 +61,9 @@
     "react-hot-loader": "^1.3.0",
     "style-loader": "^0.13.1",
     "url-loader": "^0.5.7",
-    "webpack": "^1.13.2",
-    "webpack-dev-server": "^1.15.0"
+    "webpack": "^1.14.0",
+    "webpack-dev-middleware": "^1.10.1",
+    "webpack-dev-server": "^1.15.0",
+    "webpack-hot-middleware": "^2.18.0"
   }
 }
